# feat!: 9x faster Sixel image preview