--- a/yazi-adapter/Cargo.toml
+++ b/yazi-adapter/Cargo.toml
@@ -18,9 +18,10 @@ yazi-term   = { path = "../yazi-term", version = "25.4.8" }
 ansi-to-tui = { workspace = true }
 anyhow      = { workspace = true }
 base64      = { workspace = true }
-color_quant = "1.1.0"
 crossterm   = { workspace = true }
 image       = { version = "0.25.6", default-features = false, features = [ "avif", "bmp", "dds", "exr", "ff", "gif", "hdr", "ico", "jpeg", "png", "pnm", "qoi", "tga", "webp" ] }
+palette     = { version = "0.7.6", default-features = false }
+quantette   = { version = "0.3.0", default-features = false }
 ratatui     = { workspace = true }
 scopeguard  = { workspace = true }
 tokio       = { workspace = true }