See https://github.com/sxyazi/yazi/issues/2681

Closes https://github.com/sxyazi/yazi/issues/2681

---

This PR replaces the existing neural network quantization algorithm with Wu's Color Quantizer, which offers [better performance](https://github.com/Ivordir/quantette/blob/main/docs/Benchmarks%20and%20Accuracy.md#time) while [maintaining comparable quality](https://github.com/Ivordir/quantette/blob/main/docs/Benchmarks%20and%20Accuracy.md#accuracydssim).

I've integrated the new algorithm into Yazi's Sixel encoding process, and the benchmark has shown that the speed is 9 times faster than before.

Old implementation (neural network via [`color_quant`](https://github.com/image-rs/color_quant)) – average 189.5763999ms:

```sh
time: 190.025ms
time: 191.054417ms
time: 187.696792ms
time: 188.060791ms
time: 189.381958ms
time: 191.618042ms
time: 192.383125ms
time: 190.511291ms
time: 187.8395ms
time: 187.193083ms
```

New implementation (Wu - sRGB via [`quantette`](https://github.com/Ivordir/quantette)) – average 20.9199916ms:

```sh
time: 21.226458ms
time: 20.719542ms
time: 21.207583ms
time: 20.815875ms
time: 21.071291ms
time: 21.273459ms
time: 20.811792ms
time: 20.723166ms
time: 20.52675ms
time: 20.824ms
```

Benchmark script:

```rust
// yazi-adapter/src/drivers/sixel.rs
mod tests {
  use super::*;

  #[tokio::test]
  async fn test_sixel() {
    yazi_shared::init();
    yazi_config::init().unwrap();

    let img = image::ImageReader::open("/tmp/test.jpg").unwrap().decode().unwrap();

    for _ in 0..10 {
      let img = img.clone();

      let now = std::time::Instant::now();
      _ = Sixel::encode(img).await.unwrap();
      println!("time: {:?}", now.elapsed());
    }
  }
}
```

Then run the test in release mode:

```sh
cargo test --package yazi-adapter --lib --profile release -- drivers::sixel::tests::test_sixel --exact --show-output
```

With:

![test](https://github.com/user-attachments/assets/c286b2d6-1b69-4cb0-b8a9-fda3a26116e5)

## ⚠️ Breaking changes

The `sixel_fraction` config option has been removed, as it was specific to the old neural network algorithm and is no longer needed in the new implementation.